requirements.txt
requests
pytest-xdist
pytest-ordering
pytest
allure-pytest
pyyaml
cryptography
python-dotenv
jsonpath
pytest-rerunfailures
openai